requirements.txt
# Core Django
Django==4.2.9
djangorestframework==3.14.0

# Database
mysqlclient==2.2.1

# WebSockets & Real-time
channels==4.0.0
channels-redis==4.1.0

# Authentication
djangorestframework-simplejwt==5.3.1

# CORS
django-cors-headers==4.3.1

# Environment variables
python-decouple==3.8

# Redis
redis==5.0.1

# Async support
daphne==4.0.0

# Utilities
python-dateutil==2.8.2
Pillow==10.2.0
